Changed data/model/LiquidWithAlcohol alcohol percentage to double to match the DTO.

+++ Data/Model/LiquidWithAlcohol.cs
@@ -11,7 +11,7 @@ public class LiquidWithAlcohol : LiquidProduct
     {
     }
 
-    public LiquidWithAlcohol(string name, decimal costPrice, int stockQuantity, int volumeCl, decimal salesPrice, decimal alcoholPercentage): 
+    public LiquidWithAlcohol(string name, decimal costPrice, int stockQuantity, int volumeCl, decimal salesPrice, double alcoholPercentage): 
         base(name, costPrice,stockQuantity, volumeCl, salesPrice)
     {
         AlcoholPercentage = alcoholPercentage;
